Javascript Files management
