+++ model/attraction.py
@@ -21,6 +21,7 @@ class AttractionModel:
     # === JOIN捷運站列表必須用 LEFT JOIN，不然捷運站為null的景點會被忽略
     # 創建一個 view 先把需要的 attraction 資料合併起來
 
+
     def get_attraction_data_by_page_and_keyword(page: int, keyword: str, page_size: int):
         sql_base = "SELECT * FROM taipei_day_trip.attraction_view"
         # 如果未提供關鍵字，獲取所有景點資料
@@ -40,12 +41,15 @@ class AttractionModel:
         # 如果沒有找到符合關鍵字的景點資料，回傳 None (fetchall()會回傳空列表)
         if len(results) == 0:
             return None
+
         # 因為圖片列表是字串格式，回傳前要先處理
         else:
             for result in results:
                 img_list = result["images"].split(",")
                 result["images"] = img_list
+
         print("以關鍵字獲取景點資料成功")
+
         return results
     ############################################################
 
@@ -76,4 +80,5 @@ class AttractionModel:
             # 因為圖片列表是字串格式，回傳前要先處理
             img_list = results["images"].split(",")
             results["images"] = img_list
+
             return results
