+++ pyproject.toml
@@ -15,6 +15,7 @@ python-jose = {extras = ["cryptography"], version = "^3.3.0"}
 passlib = {extras = ["bcrypt"], version = "^1.7.4"}
 requests = "^2.32.3"
 pillow = "^10.3.0"
+cloudinary = "^1.40.0"
 
 
 [build-system]
