+++ static/js/signup_login.js
@@ -1,54 +1,11 @@
+//彈出頁面相關 Module
 import { showLoginModal, hideLoginModal, togglePages } from "./view/popupModal.js";
+//送出註冊 /登入請求到後端 Module
+import { registerUser, loginUser } from "./api/signup_loginRequest.js";
 
-//顯示成功/錯誤訊息
-function showMessage(field, message, type = "success") {
-  const page = document.getElementById(`${field}`);
-  const form = document.getElementById(`${field}Form`);
-  const formLink = document.getElementById(`${field}LinkContent`);
+//顯示成功 /錯誤訊息 Module
+import { showMessage } from "./view/showMessage.js";
 
-  let oldMessage = document.querySelector(".form_message");
-
-  //如果有舊的訊息，則先移除
-  if (oldMessage) {
-    form.removeChild(oldMessage);
-  }
- 
-  let text = document.createElement("p");
-  const div = document.createElement("div");
-  const icon = document.createElement("i");
-
-  // 設定 icon 和 message 的顏色
-  switch (type) {
-    case "success":
-      icon.classList.add("fa-solid", "fa-check-circle");
-      div.classList.add("form_message", "success");
-      break;
-    case "fail":
-      icon.classList.add("fa-solid", "fa-circle-exclamation");
-      div.classList.add("form_message", "fail");
-      break;
-    case "pending":
-      icon.classList.add("fa-solid", "fa-warning");
-      div.classList.add("form_message", "pending");
-      break;
-    default:
-      console.log(`Unknown type: ${type}`);
-      break
-  }
-
-  text.textContent = message;
-
-  div.appendChild(icon);
-  div.appendChild(text);
-  form.insertBefore(div, formLink);
-
-  if (type === 'fail') {
-    page.classList.remove("shake-effect");
-    setTimeout(() => {
-      page.classList.add("shake-effect");
-    }, 100);
-  }
-}
 
 //把表單清乾淨
 function clearForm(field){
@@ -67,43 +24,6 @@ function clearForm(field){
     message.remove();
   }
 }
-
-
-//登入使用者
-async function loginUser(requestData) {
-
-  try {
-    const response = await fetch("/api/user/auth", {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify(requestData)
-    })
-
-    const result = await response.json();
-
-    if (!response.ok) {
-      // 登入失敗，顯示錯誤訊息
-      if (response.status === 400) {
-        showMessage("login", result.message, "fail");
-      }
-      //輸入格式錯誤，顯示錯誤訊息
-      if (response.status === 422) {
-        showMessage("login", "請輸入正確的電子信箱格式", "fail");
-      }
-
-      throw new Error(result.message);
-    }
-    
-    return result;
-
-  } catch (error) {
-    console.error("Error: ", error)
-  }
-  
-}
-
 
 // 送出登入表單
 async function submitLoginForm(e) {
@@ -142,41 +62,6 @@ async function submitLoginForm(e) {
     console.error("登入失敗");
   }
     
-}
-
-// 註冊使用者
-async function registerUser(requestData) {
-  try {
-    const response = await fetch("/api/user", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify(requestData)
-    })
-
-    const result = await response.json();
-
-    if (!response.ok) {
-      // 註冊失敗，顯示錯誤訊息
-      if (response.status === 400) {
-        showMessage("signup", result.message, "fail");
-      }
-
-      //輸入格式錯誤，顯示錯誤訊息
-      if (response.status === 422) {
-        showMessage("signup", "請輸入正確的電子信箱格式", "fail");
-      }
-
-      throw new Error(result.message);
-    }
-
-    return true;
-
-  } catch (error) {
-    console.error("Error: ", error)
-    return false;
-  }
 }
 
 //送出註冊表單
